FAC-950: Feat : Remove `adviceAssessmentKitPersistenceJpaAdapter` and move its functionality to `Assessment` packages
